requirements.txt
streamlit
torch
torchvision
opencv-python
pandas
Pillow